requirements.txt
astroid==2.8.3
autopep8==1.6.0
beautifulsoup4==4.10.0
build==0.7.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
flake8==4.0.1
greenlet==1.1.2
idna==3.3
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
packaging==21.0
pep517==0.12.0
pip==21.3.1
platformdirs==2.4.0
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
pylint==2.11.1
pyparsing==3.0.1
python-dotenv==0.19.1
requests==2.26.0
setuptools==57.4.0
six==1.16.0
snowballstemmer==2.1.0
soupsieve==2.2.1
SQLAlchemy==1.4.26
SQLAlchemy-Utils==0.37.9
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
wheel==0.37.0
wrapt==1.13.2
